backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.0
requests==2.31.0
cachetools==5.3.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
